requirements.txt
streamlit
gradio
torch
transformers
Pillow
requests
python-dotenv